+++ templates/sync-function-validation-module.js
@@ -311,11 +311,11 @@ function validationModule() {
       resolvedPropertyValidators.type = typeIdValidator;
     }
 
-    // Execute each of the document's property validators while ignoring the specified whitelisted properties at the root level
+    // Execute each of the document's property validators while ignoring internal document properties at the root level
     validateProperties(
       resolvedPropertyValidators,
       resolveDocConstraint(doc, oldDoc, docDefinition.allowUnknownProperties),
-      [ '_id', '_rev', '_deleted', '_revisions', '_attachments' ]);
+      true);
 
     if (doc._attachments) {
       validateAttachments();
@@ -333,7 +333,7 @@ function validationModule() {
       }
     }
 
-    function validateProperties(propertyValidators, allowUnknownProperties, whitelistedProperties) {
+    function validateProperties(propertyValidators, allowUnknownProperties, ignoreInternalProperties) {
       var currentItemEntry = itemStack[itemStack.length - 1];
       var objectValue = currentItemEntry.itemValue;
       var oldObjectValue = currentItemEntry.oldItemValue;
@@ -369,7 +369,7 @@ function validationModule() {
       // Verify there are no unsupported properties in the object
       if (!allowUnknownProperties) {
         for (var propertyName in objectValue) {
-          if (whitelistedProperties && whitelistedProperties.indexOf(propertyName) >= 0) {
+          if (ignoreInternalProperties && propertyName.indexOf('_') === 0) {
             // These properties are special cases that should always be allowed - generally only applied at the root level of the document
             continue;
           }
@@ -463,11 +463,10 @@ function validationModule() {
 
         switch (validatorType) {
           case 'string':
-            var regexPattern = resolveValidationConstraint(validator.regexPattern);
             if (typeof itemValue !== 'string') {
               validationErrors.push('item "' + buildItemPath(itemStack) + '" must be a string');
-            } else if (regexPattern && !regexPattern.test(itemValue)) {
-              validationErrors.push('item "' + buildItemPath(itemStack) + '" must conform to expected format ' + regexPattern);
+            } else {
+              validateString(validator);
             }
             break;
           case 'integer':
@@ -487,17 +486,17 @@ function validationModule() {
             break;
           case 'datetime':
             if (typeof itemValue !== 'string' || !isIso8601DateTimeString(itemValue)) {
-              validationErrors.push('item "' + buildItemPath(itemStack) + '" must be an ISO 8601 date string with optional time and time zone components');
+              validationErrors.push('item "' + buildItemPath(itemStack) + '" must be an ECMAScript simplified ISO 8601 date string with optional time and time zone components');
             }
             break;
           case 'date':
             if (typeof itemValue !== 'string' || !isIso8601DateString(itemValue)) {
-              validationErrors.push('item "' + buildItemPath(itemStack) + '" must be an ISO 8601 date string with no time or time zone components');
+              validationErrors.push('item "' + buildItemPath(itemStack) + '" must be an ECMAScript simplified ISO 8601 date string with no time or time zone components');
             }
             break;
           case 'time':
             if (typeof itemValue !== 'string' || !isIso8601TimeString(itemValue)) {
-              validationErrors.push('item "' + buildItemPath(itemStack) + '" must be an ISO 8601 time string with no date or time zone components');
+              validationErrors.push('item "' + buildItemPath(itemStack) + '" must be an ECMAScript simplified ISO 8601 time string with no date or time zone components');
             }
             break;
           case 'timezone':
@@ -510,12 +509,12 @@ function validationModule() {
             if (!(enumPredefinedValues instanceof Array)) {
               validationErrors.push('item "' + buildItemPath(itemStack) + '" belongs to an enum that has no predefined values');
             } else if (enumPredefinedValues.indexOf(itemValue) < 0) {
-              validationErrors.push('item "' + buildItemPath(itemStack) + '" must be one of the predefined values: ' + enumPredefinedValues.toString());
+              validationErrors.push('item "' + buildItemPath(itemStack) + '" must be one of the predefined values: ' + enumPredefinedValues.join(','));
             }
             break;
           case 'uuid':
             if (!isUuid(itemValue)) {
-              validationErrors.push('item "' + buildItemPath(itemStack) + '" is not a valid UUID');
+              validationErrors.push('item "' + buildItemPath(itemStack) + '" must be a UUID string');
             }
             break;
           case 'object':
@@ -541,7 +540,7 @@ function validationModule() {
         }
       } else if (resolveValidationConstraint(validator.required)) {
         // The item has no value (either it's null or undefined), but the validator indicates it is required
-        validationErrors.push('required item "' + buildItemPath(itemStack) + '" is missing');
+        validationErrors.push('item "' + buildItemPath(itemStack) + '" must not be null or missing');
       } else if (resolveValidationConstraint(validator.mustNotBeMissing) && typeof itemValue === 'undefined') {
         // The item is missing (i.e. it's undefined), but the validator indicates it must not be
         validationErrors.push('item "' + buildItemPath(itemStack) + '" must not be missing');
@@ -553,6 +552,29 @@ function validationModule() {
 
     function hasNoValue(value, treatNullAsUndefined) {
       return treatNullAsUndefined ? isValueNullOrUndefined(value) : typeof value === 'undefined';
+    }
+
+    function validateString(validator) {
+      var currentItemEntry = itemStack[itemStack.length - 1];
+      var itemValue = currentItemEntry.itemValue;
+
+      var regexPattern = resolveValidationConstraint(validator.regexPattern);
+      if (regexPattern && !regexPattern.test(itemValue)) {
+        validationErrors.push('item "' + buildItemPath(itemStack) + '" must conform to expected format ' + regexPattern);
+      }
+
+      var mustBeTrimmed = resolveValidationConstraint(validator.mustBeTrimmed);
+      if (mustBeTrimmed && isStringUntrimmed(itemValue)) {
+        validationErrors.push('item "' + buildItemPath(itemStack) + '" must not have any leading or trailing whitespace');
+      }
+    }
+
+    function isStringUntrimmed(value) {
+      if (isValueNullOrUndefined(value)) {
+        return false;
+      } else {
+        return value !== value.trim();
+      }
     }
 
     function validateImmutable(onlyEnforceIfHasValue, treatNullAsUndefined) {
@@ -578,7 +600,7 @@ function validationModule() {
         }
 
         if (!constraintSatisfied) {
-          validationErrors.push('value of item "' + buildItemPath(itemStack) + '" may not be modified');
+          validationErrors.push('item "' + buildItemPath(itemStack) + '" cannot be modified');
         }
       }
     }
@@ -724,11 +746,11 @@ function validationModule() {
               validationErrors.push('hashtable key "' + fullKeyPath + '" is not a string');
             } else {
               if (resolveValidationConstraint(keyValidator.mustNotBeEmpty) && elementKey.length < 1) {
-                validationErrors.push('empty hashtable key in item "' + buildItemPath(itemStack) + '" is not allowed');
+                validationErrors.push('hashtable "' + buildItemPath(itemStack) + '" must not have an empty key');
               }
               var regexPattern = resolveValidationConstraint(keyValidator.regexPattern);
               if (regexPattern && !regexPattern.test(elementKey)) {
-                validationErrors.push('hashtable key "' + fullKeyPath + '" does not conform to expected format ' + regexPattern);
+                validationErrors.push('hashtable key "' + fullKeyPath + '" must conform to expected format ' + regexPattern);
               }
             }
           }
@@ -768,7 +790,7 @@ function validationModule() {
       var itemValue = currentItemEntry.itemValue;
 
       if (typeof itemValue !== 'string') {
-        validationErrors.push('attachment reference "' + buildItemPath(itemStack) + '" must be a string');
+        validationErrors.push('item "' + buildItemPath(itemStack) + '" must be an attachment reference string');
       } else {
         attachmentReferenceValidators[itemValue] = validator;
 
@@ -861,11 +883,11 @@ function validationModule() {
       }
 
       if (isInteger(maximumTotalAttachmentSize) && totalSize > maximumTotalAttachmentSize) {
-        validationErrors.push('the total size of all attachments must not exceed ' + maximumTotalAttachmentSize + ' bytes');
+        validationErrors.push('documents of this type must not have a combined attachment size greater than ' + maximumTotalAttachmentSize + ' bytes');
       }
 
       if (isInteger(maximumAttachmentCount) && attachmentCount > maximumAttachmentCount) {
-        validationErrors.push('the total number of attachments must not exceed ' + maximumAttachmentCount);
+        validationErrors.push('documents of this type must not have more than ' + maximumAttachmentCount + ' attachments');
       }
 
       if (!resolveDocConstraint(doc, oldDoc, docDefinition.allowAttachments) && attachmentCount > 0) {