+++ CHANGELOG.md
@@ -5,8 +5,13 @@ This project adheres to [Semantic Versioning](http://semver.org/). All notable c
 ### Added
 - [#43](https://github.com/Kashoo/synctos/issues/43): Tool to validate structure and semantics of a document definitions file
 - [#189](https://github.com/Kashoo/synctos/issues/189): Automatically create the output sync function file directory if it does not exist
+- [#207](https://github.com/Kashoo/synctos/issues/207): Ignore all top-level document properties that start with an underscore
+- [#204](https://github.com/Kashoo/synctos/issues/204): Constraint that requires string values to be trimmed
 - [#180](https://github.com/Kashoo/synctos/issues/180): Data validation type for time of day
 - [#202](https://github.com/Kashoo/synctos/issues/202): Time zone data validation type
+
+### Changed
+- [#212](https://github.com/Kashoo/synctos/issues/212): Improve document validation error messages
 
 ### Fixed
 - [#190](https://github.com/Kashoo/synctos/issues/190): JavaScript error when mustEqual constraint is violated